# src/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
...
paths:
  /api/users/edit-profile:
    put:
      tags:
        - 5. User Actions (Profile)
      summary: Cập nhật thông tin cá nhân
      description: 'Người dùng tự cập nhật thông tin (tên, ngày sinh...). Yêu cầu xác thực.'
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserResponseDto'
  '/api/users/{userId}/image':
    post:
      tags:
        - 5. User Actions (Profile)
      summary: Tải lên ảnh đại diện
      description: Tải lên ảnh đại diện (avatar) cho người dùng.
      operationId: uploadUserImage
      parameters:
        - name: userId
          in: path
          description: ID của người dùng
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: File ảnh
                  format: binary
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUploadImageResponse'
...
  '/api/users/{id}/change-password':
    patch:
      tags:
        - 5. User Actions (Profile)
      summary: Đổi mật khẩu
      description: >-
        Người dùng tự đổi mật khẩu của mình (yêu cầu mật khẩu cũ và mới). Yêu
        cầu xác thực (USER/ADMIN/TEACHER).
      operationId: changePassword
      parameters:
        - name: id
          in: path
          description: ID của người dùng
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          description: Đổi mật khẩu thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
        '400':
          description: Mật khẩu cũ không chính xác hoặc mật khẩu mới không khớp
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
        '403':
          description: Không có quyền đổi mật khẩu cho user này
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseString'
...
  '/api/teacher/courses/{courseId}':
    get:
      tags:
        - 4. Teacher - Course Creation
      summary: '[TEACHER] Lấy chi tiết khóa học để sửa (Load nháp)'
      description: >-
        Lấy toàn bộ cây thông tin lồng nhau (Section, Lesson, Exam...) của một
        khóa học để Teacher tiếp tục chỉnh sửa. Yêu cầu quyền TEACHER/ADMIN.
      operationId: getMyCourseForEdit
      parameters:
        - name: courseId
          in: path
          description: ID của khóa học cần sửa
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCourseDetailDto'
        '403':
          description: Không có quyền (không phải chủ khóa học)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCourseDetailDto'
        '404':
          description: Không tìm thấy khóa học
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCourseDetailDto'
  /api/public/courses:
    get:
      tags:
        - 3. Public Courses
      summary: '[PUBLIC] Lấy danh sách khóa học (Trang chủ)'
      description: >-
        Lấy danh sách các khóa học đã được 'APPROVED' (phê duyệt), có phân
        trang. (API này nên được sửa để trả về PagedResponse)
      operationId: getAllPublicCourses
      parameters:
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: 'Thành công, trả về danh sách khóa học'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePagePublicCourseListDto'
  '/api/public/courses/{id}':
    get:
      tags:
        - 3. Public Courses
      summary: '[PUBLIC] Lấy chi tiết 1 khóa học (để xem)'
      description: >-
        Lấy thông tin chi tiết của 1 khóa học đã được 'APPROVED'. Chỉ trả về
        thông tin công khai (tiêu đề bài giảng), không bao gồm nội dung.
      operationId: getPublicCourseDetails
      parameters:
        - name: id
          in: path
          description: ID của khóa học cần xem
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Thành công, trả về chi tiết khóa học'
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePublicCourseDetailDto'
        '404':
          description: Không tìm thấy khóa học (hoặc khóa học chưa được duyệt)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePublicCourseDetailDto'
...
  '/api/admin/courses/{id}':
    get:
      tags:
        - 1. Admin Management
      summary: '[ADMIN] Lấy chi tiết 1 khóa học'
      description: >-
        Lấy toàn bộ cây thông tin lồng nhau của một khóa học. Yêu cầu quyền
        ADMIN.
      operationId: getCourseDetails
      parameters:
        - name: id
          in: path
          description: ID của khóa học cần xem
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Thành công
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCourseDetailDto'
        '404':
          description: Không tìm thấy khóa học
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAdminCourseDetailDto'
components:
  schemas:
    UpdateProfileRequest:
      required:
        - firstname
      type: object
      properties:
        firstname:
          maxLength: 20
          minLength: 0
          type: string
        lastname:
          type: string
        image:
          type: string
        date_of_birth:
          type: string
          format: date
    ApiResponseUserResponseDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          $ref: '#/components/schemas/UserResponseDto'
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        dateOfBirth:
          type: string
          format: date
        role:
          type: string
          enum:
            - ADMIN
            - TEACHER
            - STUDENT
        image:
          type: string
        active:
          type: boolean
    ApiResponseUploadImageResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          $ref: '#/components/schemas/UploadImageResponse'
    UploadImageResponse:
      type: object
      properties:
        imageUrl:
          type: string
...
    ApiResponseString:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          type: string
...
    ChangePasswordRequest:
      required:
        - newPassword
        - oldPassword
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
...
    AdminAnswerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        content:
          type: string
        isCorrect:
          type: boolean
    AdminAuthorDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
    AdminCourseDetailDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - PENDING_APPROVAL
            - ACTIVE
            - REJECTED
            - ARCHIVED
        image:
          type: string
        objectives:
          type: string
        targetAudience:
          type: string
        category:
          type: string
          enum:
            - COMPUTER_SCIENCE
            - MATHEMATICS
            - PHYSICS
            - BUSINESS_ADMINISTRATION
            - LITERATURE
        author:
          $ref: '#/components/schemas/AdminAuthorDto'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/AdminSectionDto'
    AdminExamDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        examType:
          type: string
          enum:
            - QUIZ
            - ASSIGNMENT
            - FINAL_EXAM
        questions:
          type: array
          items:
            $ref: '#/components/schemas/AdminQuestionDto'
    AdminLessonDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        lessonType:
          type: string
          enum:
            - VIDEO
            - ARTICLE
        videoUrl:
          type: string
        duration:
          type: integer
          format: int64
        articleContent:
          type: string
        order:
          type: integer
          format: int32
    AdminQuestionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        questionType:
          type: string
          enum:
            - SINGLE_CHOICE
            - MULTIPLE_CHOICE
            - ESSAY
        content:
          type: string
        point:
          type: integer
          format: int32
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AdminAnswerDto'
    AdminSectionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        order:
          type: integer
          format: int32
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/AdminLessonDto'
        exams:
          type: array
          items:
            $ref: '#/components/schemas/AdminExamDto'
    ApiResponseAdminCourseDetailDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          $ref: '#/components/schemas/AdminCourseDetailDto'
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    ApiResponsePagePublicCourseListDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          $ref: '#/components/schemas/PagePublicCourseListDto'
    PagePublicCourseListDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/PublicCourseListDto'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    PublicAuthorDto:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        image:
          type: string
    PublicCourseListDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        image:
          type: string
        category:
          type: string
          enum:
            - COMPUTER_SCIENCE
            - MATHEMATICS
            - PHYSICS
            - BUSINESS_ADMINISTRATION
            - LITERATURE
        author:
          $ref: '#/components/schemas/PublicAuthorDto'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    ApiResponsePublicCourseDetailDto:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        errorCode:
          type: string
        data:
          $ref: '#/components/schemas/PublicCourseDetailDto'
    PublicCourseDetailDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        image:
          type: string
        objectives:
          type: string
        targetAudience:
          type: string
        category:
          type: string
          enum:
            - COMPUTER_SCIENCE
            - MATHEMATICS
            - PHYSICS
            - BUSINESS_ADMINISTRATION
            - LITERATURE
        author:
          $ref: '#/components/schemas/PublicAuthorDto'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/PublicSectionDto'
    PublicLessonDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        duration:
          type: integer
          format: int64
        order:
          type: integer
          format: int32
        lessonType:
          type: string
          enum:
            - VIDEO
            - ARTICLE
    PublicSectionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        order:
          type: integer
          format: int32
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/PublicLessonDto'
...